requirements.txt
# Core ML libraries
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
xgboost==1.7.6
lightgbm==4.0.0
catboost==1.2.2

# Deep Learning
tensorflow==2.13.0
torch==2.0.1

# Data visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Data validation and profiling
great-expectations==0.17.23
pandas-profiling==3.6.6

# Model tracking and experiment management
mlflow==2.6.0
wandb==0.15.8

# Feature engineering
feature-engine==1.6.2
category-encoders==2.6.1

# Hyperparameter optimization
optuna==3.2.0
hyperopt==0.2.7

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1

# Code quality and security
black==23.7.0
flake8==6.0.0
isort==5.12.0
bandit==1.7.5
safety==2.3.5

# Configuration management
pyyaml==6.0.1
python-dotenv==1.0.0

# API and web framework
fastapi==0.100.1
uvicorn==0.23.2
pydantic==2.1.1

# Database
sqlalchemy==2.0.19
psycopg2-binary==2.9.7

# Monitoring and logging
structlog==23.1.0
prometheus-client==0.17.1

# Utilities
tqdm==4.65.0
joblib==1.3.2
scipy==1.11.1
kaggle==1.6.14
